requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.122.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
pyahocorasick==2.0.0
tqdm==4.67.1

# Local LLM
ollama==0.6.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Development utilities
aiofiles==24.1.0
python-docx==1.1.2
